test/unit tests

+++ restaurant-service/src/main/java/com/fooddelivery/restaurant/dto/MenuItemPriceDto.java
@@ -1,14 +1,9 @@
 package com.fooddelivery.restaurant.dto;
 
 import lombok.Builder;
-import java.math.BigDecimal;
 
 /** Lightweight DTO returned to Order Service for price validation */
 @Builder
 public record MenuItemPriceDto(
-    Long id,
-    String name,
-    BigDecimal price,
-    Long restaurantId,
-    boolean available
-) {}
+    Long id, String name, Integer price, Long restaurantId, boolean available) {}
+
